main: Add secure server which responses securely
This server has the same logic as the regualr server, except the
secure mechanism.

The private and the public keys weren't commited from security reasons,
so in order to start the server properly, place the following files in
their assigned locations to start the server successfully:
1. private key
  - file name: private.key
  - location: inside "tls" folder
2. public key
  - file name: public.crt
  - location: inside "tls" folder

This is an addition in infrastructure level, means that there are no
actual calls that do real things right now, and if there are, they
might have unexpected behavior.